requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
google-genai
python-dotenv
psycopg2-binary
PyJWT[crypto]
httpx
anthropic
pydantic
